[clear]frontのリファクタリング
front部分のリファクタリング。atomic designに倣ってコンポーネント化
