+++ front/components/Atoms/SiteTitle.tsx
@@ -1,0 +1,13 @@
+const SiteTitle = () => {
+  return (
+    <h1
+      className={
+        'font-serif text-5xl text-center font-bold text-gray-900 max-md:text-3xl'
+      }
+    >
+      Travel Weather
+    </h1>
+  );
+};
+
+export default SiteTitle;
