+++ front/lib/type/dataType.ts
@@ -1,0 +1,15 @@
+type data = {
+  id: number;
+  city_id: number;
+  date_time: string;
+  temp: number;
+  temp_max: number;
+  temp_min: number;
+  humidity: number;
+  description: string;
+  alert: string;
+  icon: string;
+  created_at: string;
+  updated_at: string;
+  weather: string;
+};
