+++ front/lib/type/detailType.ts
@@ -1,0 +1,5 @@
+type detail = {
+  color: string;
+  title: string;
+  content: string;
+};
